[TEST] Added the `skip` configuration for `warnings` check in the "Indices Analyze" tests
As part of b6537d719d840dc987a1fb2df4b48bf260e53371, a check for the `Warnings` headers
has been added. Since the feature might be temporarily or permanently not supported by all clients
(as is the case with eg. the `headers` feature), we should allow them to skip the tests.

Related: #20246, #20686

CC @johtani